requirements.txt
selenium==4.9.0
pytest==7.3.0
allure-pytest==2.13.1
allure-python-commons==2.13.1
webdriver-manager==3.8.6